refactor: move RecordStat from models to stats package
RecordStat uses sync/atomic, which violates the models/ package
contract of being pure data structures. Move it to stats/ where
concurrency primitives belong.

- Create stats/record_stat.go with RecordStat and atomic methods
- Remove RecordStat and sync/atomic import from models/models.go
- Update record/record.go and proxy/proxy.go to import from stats
- Update proxy/proxy_test.go to import from stats
